HBASE-25781 Add isCacheBlocks option to RowCounter

+++ hbase-common/src/main/java/org/apache/hadoop/hbase/util/AbstractHBaseTool.java
@@ -244,7 +244,7 @@ public abstract class AbstractHBaseTool implements Tool {
     }
   }
 
-  public long getOptionAsLong(CommandLine cmd, String opt, int defaultValue) {
+  public long getOptionAsLong(CommandLine cmd, String opt, long defaultValue) {
     return getOptionAsLong(cmd, opt, defaultValue, 10);
   }
